# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Swagger documentation for the Gin-Golang project.",
        "title": "Gin-Golang API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/attendees/events/{userId}": {
            "get": {
                "description": "Lấy danh sách tất cả sự kiện đã tham gia của người tham dự này",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendees"
                ],
                "summary": "Trả toạn bộ sự kiện mà người tham dự đã tham gia",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id của user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    }
                }
            }
        },
        "/attendees/register": {
            "post": {
                "description": "Chèn dữ liệu tham dự của người dùng với sự kiện",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendees"
                ],
                "summary": "Đăng kí tham dự sự kiên",
                "parameters": [
                    {
                        "description": "Thông tin đăng ký tham dự",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAttendeeDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Attendee"
                        }
                    }
                }
            }
        },
        "/attendees/users/{eventId}": {
            "get": {
                "description": "Lấy danh sách tất cả người tham dự sự kiện này",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendees"
                ],
                "summary": "Trả toàn bộ người tham dự sự kiện",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id của event",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                }
            }
        },
        "/attendees/{eventId}/{userId}": {
            "delete": {
                "description": "Xóa dự liệu tham dự sự kiện của người tham dự",
                "tags": [
                    "Attendees"
                ],
                "summary": "Xóa dự liệu tham dự sự kiện",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id của event",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id của user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/events": {
            "get": {
                "description": "Lấy danh sách tất cả events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Trả toàn bộ events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Thêm event mới vào database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Tạo event mới",
                "parameters": [
                    {
                        "description": "Thông tin event vừa thêm",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateEventDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    }
                }
            }
        },
        "/events/images": {
            "post": {
                "description": "Upload mỗi file \u003c= 10MB. Nếu có 1 file sai định dạng → KHÔNG LƯU BẤT KÌ FILE NÀO",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Upload ảnh cho event **MULTIPLE FILE**",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của event",
                        "name": "eventId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Danh sách file ảnh",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "description": "Trích xuất event cụ thể bằng ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Lấy event theo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id của event",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    }
                }
            },
            "put": {
                "description": "Cập nhập event đã tồn tại bằng ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Cập nhập event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id của event",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dữ liệu event sau cập nhập",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEventDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    }
                }
            },
            "delete": {
                "description": "Xóa event theo Id",
                "tags": [
                    "Events"
                ],
                "summary": "Xóa event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id của event",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateAttendeeDto": {
            "type": "object",
            "required": [
                "eventId",
                "userId"
            ],
            "properties": {
                "eventId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateEventDto": {
            "type": "object",
            "required": [
                "date",
                "description",
                "location",
                "name",
                "ownerId"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "minLength": 10
                },
                "location": {
                    "type": "string",
                    "minLength": 3
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "ownerId": {
                    "type": "integer"
                }
            }
        },
        "dto.DataResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "dto.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateEventDto": {
            "type": "object",
            "required": [
                "date",
                "description",
                "location",
                "name"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "minLength": 10
                },
                "location": {
                    "type": "string",
                    "minLength": 3
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
        "models.Attendee": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "models.Event": {
            "type": "object",
            "required": [
                "date",
                "description",
                "location",
                "name",
                "ownerId"
            ],
            "properties": {
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "minLength": 10
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string",
                    "minLength": 3
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "ownerId": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}